pyproject.toml
[tool.poetry]
name = "webgpt"
version = "0.1.0"
description = ""
authors = ["Jonatan Tierno <jonatan.tiernoalvite@telefonica.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.0"
Jinja2 = "^3.1.2"
Flask = "^2.0.2"
python-dotenv = "^1.0.0"
ffmpeg-python = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
